feat: 🎸 outsource websocket logic into own file
and add waiting room websocket connection
