+++ src/components/app/app.tsx
@@ -4,11 +4,13 @@ import * as ReactDOM from 'react-dom';
 import { useState } from 'react';
 import { Router, Switch, Route } from 'react-router-dom';
 import { createBrowserHistory } from 'history';
-import { AuthenticatedAppContainer } from './AuthenticatedApp';
+import { AuthenticatedApp } from './AuthenticatedApp';
 import { Registration } from '../registration/Registration';
 import { LoginLoader } from './LoginLoader';
 import '../../resources/styles/styles';
 import { WaitingRoomLoader } from '../waitingRoom/WaitingRoomLoader';
+import { ContextProvider } from '../../globalState/state';
+import { WebsocketHandler } from './WebsocketHandler';
 
 export const history = createBrowserHistory();
 
@@ -17,46 +19,71 @@ const App = () => {
 	// optional resort name. Since resort names are dynamic, we have
 	// to find out if the provided path is a resort name. If not, we
 	// use the authenticated app as a catch-all fallback.
-	const [hasUnmatchedResortLogin, setHasUnmatchedResortLogin] = useState(
-		false
-	);
+	const [
+		hasUnmatchedLoginConsultingType,
+		setHasUnmatchedLoginConsultingType
+	] = useState(false);
 	const [
 		hasUnmatchedAnonyomousConversation,
 		setHasUnmatchedAnonyomousConversation
 	] = useState(false);
-
-	const handleUnmatchResortLogin = () => {
-		setHasUnmatchedResortLogin(true);
-	};
-
-	const handleUnmatchAnonyomousConversation = () => {
-		setHasUnmatchedAnonyomousConversation(true);
-	};
+	const [
+		hasUnmatchedRegistrationConsultingType,
+		setHasUnmatchedRegistrationConsultingType
+	] = useState(false);
+	const [startWebsocket, setStartWebsocket] = useState<boolean>(false);
+	const [disconnectWebsocket, setDisconnectWebsocket] = useState<boolean>(
+		false
+	);
 
 	return (
 		<Router history={history}>
-			<Switch>
-				<Route path="/:consultingTypeSlug/registration">
-					<Registration />
-				</Route>
-				{!hasUnmatchedAnonyomousConversation && (
-					<Route path="/:consultingTypeSlug/warteraum">
-						<WaitingRoomLoader
-							handleUnmatch={handleUnmatchAnonyomousConversation}
-						/>
-					</Route>
+			<ContextProvider>
+				{startWebsocket && (
+					<WebsocketHandler disconnect={disconnectWebsocket} />
 				)}
-				{!hasUnmatchedResortLogin && (
-					<Route path="/:consultingTypeSlug">
-						<LoginLoader handleUnmatch={handleUnmatchResortLogin} />
-					</Route>
-				)}
-				<AuthenticatedAppContainer />
-			</Switch>
+				<Switch>
+					{!hasUnmatchedRegistrationConsultingType && (
+						<Route path="/:consultingTypeSlug/registration">
+							<Registration
+								handleUnmatch={() =>
+									setHasUnmatchedRegistrationConsultingType(
+										true
+									)
+								}
+							/>
+						</Route>
+					)}
+					{!hasUnmatchedAnonyomousConversation && (
+						<Route path="/:consultingTypeSlug/warteraum">
+							<WaitingRoomLoader
+								handleUnmatch={() =>
+									setHasUnmatchedAnonyomousConversation(true)
+								}
+								onAnonymousRegistration={() =>
+									setStartWebsocket(true)
+								}
+							/>
+						</Route>
+					)}
+					{!hasUnmatchedLoginConsultingType && (
+						<Route path="/:consultingTypeSlug">
+							<LoginLoader
+								handleUnmatch={() =>
+									setHasUnmatchedLoginConsultingType(true)
+								}
+							/>
+						</Route>
+					)}
+					<AuthenticatedApp
+						onAppReady={() => setStartWebsocket(true)}
+						onLogout={() => setDisconnectWebsocket(true)}
+					/>
+				</Switch>
+			</ContextProvider>
 		</Router>
 	);
 };
-
 export const initApp = () => {
 	ReactDOM.render(<App />, document.getElementById('appRoot'));
 };
